Added isRecycled to bailing if statement
In Android 8, Bitmap.getBitmapBytes responds differently than before, returning a 0 for the size if the image is recycled (where as before it would still try to calculate the size of the image). This is causing trimToSize to throw an IllegalStateException error. To resolve this, we can add to a bailing if statement inside of LruCache.set to see if the image is recycled.